Replaced floats with doubles in vertex structs

+++ src/core/vertex.h
@@ -6,8 +6,8 @@ typedef std::vector<unsigned int> neighbors_vector;
 typedef struct uvertex {
     //unsigned int id;
     neighbors_vector *neighbors;
-    float state;
-    float state_temp;
+    double state;
+    double state_temp;
 
 };
 
@@ -15,8 +15,8 @@ typedef struct dvertex {
     //unsigned int id;
     neighbors_vector *in_neighbors;
     neighbors_vector *out_neighbors;
-    float state;
-    float state_temp;
+    double state;
+    double state_temp;
 };
 
 #endif //KAPPA_VERTEX_H
